+++ js/SceneGraph_StateMachine.js
@@ -1,0 +1,35 @@
+//---------------------------------------------------------------Finite State Machine for controls -------------------------------------------------------------------------------
+export default class FiniteStateMachine {
+    constructor() {
+      this._states = {};
+      this._currentState = null;
+    }
+  
+    _AddState(name, type) {
+      this._states[name] = type;
+    }
+  
+    SetState(name) {
+      const prevState = this._currentState;
+      
+      if (prevState) {
+        if (prevState.Name == name) {
+          return;
+        }
+        prevState.Exit();
+      }
+  
+      const state = new this._states[name](this);
+  
+      this._currentState = state;
+      state.Enter(prevState);
+    }
+  
+    Update(timeElapsed, input) {
+      if (this._currentState) {
+        this._currentState.Update(timeElapsed, input);
+      }
+    }
+  };
+  
+
