+++ js/main.js
@@ -2,705 +2,10 @@
 import * as THREE from 'https://cdn.jsdelivr.net/npm/three@0.118/build/three.module.js';
 
 import {GLTFLoader} from 'https://cdn.jsdelivr.net/npm/three@0.118.1/examples/jsm/loaders/GLTFLoader.js';
-import {FBXLoader} from 'https://cdn.jsdelivr.net/npm/three@0.118.1/examples/jsm/loaders/FBXLoader.js';
 import {OrbitControls} from 'https://cdn.jsdelivr.net/npm/three@0.118/examples/jsm/controls/OrbitControls.js';
 
 
-class BasicCharacterControllerProxy {
-  constructor(animations) {
-    this._animations = animations;
-  }
-
-  get animations() {
-    return this._animations;
-  }
-};
-
-//-------------------------------------------------------------- Listen for input -------------------------------------------------------------------------------------------------
-class BasicCharacterControllerInput {
-  constructor() {
-    this._Init();    
-  }
-
-  _Init() {
-    this._keys = {
-      forward1: false,
-      backward1: false,
-      left1: false,
-      right1: false,
-      space: false,
-      m: false,
-      forward2: false,
-      backward2: false,
-      left2: false,
-      right2: false,
-      shift: false,
-      enter: false,
-    };
-    document.addEventListener('keydown', (e) => this._onKeyDown(e), false);
-    document.addEventListener('keyup', (e) => this._onKeyUp(e), false);
-  }
-
-  _onKeyDown(event) {
-    switch (event.keyCode) {
-      case 87: // w
-        this._keys.forward1 = true;
-        break;
-      case 65: // a
-        this._keys.left1 = true;
-        break;
-      case 83: // s
-        this._keys.backward1 = true;
-        break;
-      case 68: // d
-        this._keys.right1 = true;
-        break;
-      case 32: // SPACE
-        this._keys.space = true;
-        break;
-      case 77: // m
-        this._keys.m = true;
-        break;
-        case 87: // up arrow
-        this._keys.forward2 = true;
-        break;
-      case 65: // left arrow
-        this._keys.left2 = true;
-        break;
-      case 83: // down arrow
-        this._keys.backward2 = true;
-        break;
-      case 68: // right arrow
-        this._keys.right2 = true;
-        break;
-      case 32: // shift
-        this._keys.shift = true;
-        break;
-      case 77: // enter
-        this._keys.enter = true;
-        break;
-    }
-  }
-
-  _onKeyUp(event) {
-    switch(event.keyCode) {
-      case 87: // w
-        this._keys.forward1 = false;
-        break;
-      case 65: // a
-        this._keys.left1 = false;
-        break;
-      case 83: // s
-        this._keys.backward1 = false;
-        break;
-      case 68: // d
-        this._keys.right1 = false;
-        break;
-      case 32: // SPACE
-        this._keys.space = false;
-        break;
-      case 77: // m
-        this._keys.m = false;
-        break;
-        case 87: // up arrow
-        this._keys.forward2 = false;
-        break;
-      case 65: // left arrow
-        this._keys.left2 = false;
-        break;
-      case 83: // down arrow
-        this._keys.backward2 = false;
-        break;
-      case 68: // right arrow
-        this._keys.right2 = false;
-        break;
-      case 32: // shift
-        this._keys.shift = false;
-        break;
-      case 77: // enter
-        this._keys.enter = false;
-        break;
-    }
-  }
-};
-
-
-
-//---------------------------------------------------------------Finite State Machine for controls -------------------------------------------------------------------------------
-class FiniteStateMachine {
-  constructor() {
-    this._states = {};
-    this._currentState = null;
-  }
-
-  _AddState(name, type) {
-    this._states[name] = type;
-  }
-
-  SetState(name) {
-    const prevState = this._currentState;
-    
-    if (prevState) {
-      if (prevState.Name == name) {
-        return;
-      }
-      prevState.Exit();
-    }
-
-    const state = new this._states[name](this);
-
-    this._currentState = state;
-    state.Enter(prevState);
-  }
-
-  Update(timeElapsed, input) {
-    if (this._currentState) {
-      this._currentState.Update(timeElapsed, input);
-    }
-  }
-};
-
-
-class CharacterFSM extends FiniteStateMachine {
-  constructor(proxy) {
-    super();
-    this._proxy = proxy;
-    this._Init();
-  }
-
-  _Init() {
-    this._AddState('Idle1', IdleState1);
-    this._AddState('Punch1', PunchState1);
-    this._AddState('Run1', RunState1);
-    this._AddState('Jump1', JumpState1);
-    this._AddState('Idle2', IdleState2);
-    this._AddState('Punch2', PunchState2);
-    this._AddState('Run2', RunState2);
-    this._AddState('Jump2', JumpState2);
-  }
-};
-
-
-class State {
-  constructor(parent) {
-    this._parent = parent;
-  }
-
-  Enter() {}
-  Exit() {}
-  Update() {}
-};
-
-
-class JumpState1 extends State {
-  constructor(parent) {
-    super(parent);
-
-    this._FinishedCallback = () => {
-      this._Finished();
-    }
-  }
-
-  get Name() {
-    return 'Jump1';
-  }
-
-  Enter(prevState) {
-    const curAction = this._parent._proxy._animations['Jump1'].action;
-    const mixer = curAction.getMixer();
-    mixer.addEventListener('finished', this._FinishedCallback);
-
-    if (prevState) {
-      const prevAction = this._parent._proxy._animations[prevState.Name].action;
-
-      curAction.reset();  
-      curAction.setLoop(THREE.LoopOnce, 1);
-      curAction.clampWhenFinished = true;
-      curAction.crossFadeFrom(prevAction, 0.2, true);
-      curAction.play();
-    } else {
-      curAction.play();
-    }
-  }
-
-  _Finished() {
-    this._Cleanup();
-    this._parent.SetState('Idle1');
-  }
-
-  _Cleanup() {
-    const action = this._parent._proxy._animations['Jump1'].action;
-    
-    action.getMixer().removeEventListener('finished', this._CleanupCallback);
-  }
-
-  Exit() {
-    this._Cleanup();
-  }
-
-  Update(_) {
-  }
-};
-
-
-class RunState1 extends State {
-  constructor(parent) {
-    super(parent);
-  }
-
-  get Name() {
-    return 'Run1';
-  }
-
-  Enter(prevState) {
-    const curAction = this._parent._proxy._animations['Run1'].action;
-    if (prevState) {
-      const prevAction = this._parent._proxy._animations[prevState.Name].action;
-
-      curAction.enabled = true;
-      const ratio = curAction.getClip().duration / prevAction.getClip().duration;
-      curAction.time = prevAction.time * ratio;
-      curAction.crossFadeFrom(prevAction, 0.5, true);
-      curAction.play();
-    }
-  }
-
-  Exit() {
-  }
-
-  Update(timeElapsed, input) {
-    if (input._keys.forward1 || input._keys.backward1) {
-      if (input._keys.space) {
-        this._parent.SetState('Jump1');
-      }
-      if (input._keys.m) {
-        this._parent.SetState('Punch1');
-      }
-      return;
-    }
-
-    this._parent.SetState('Idle1');
-  }
-};
-
-
-class PunchState1 extends State {
-  constructor(parent) {
-    super(parent);
-  }
-
-  get Name() {
-    return 'Punch1';
-  }
-
-  Enter(prevState) {
-    const curAction = this._parent._proxy._animations['Punch1'].action;
-    if (prevState) {
-      const prevAction = this._parent._proxy._animations[prevState.Name].action;
-      curAction.enabled = true;
-      const ratio = curAction.getClip().duration / prevAction.getClip().duration;
-      curAction.time = 60;
-      curAction.crossFadeFrom(prevAction, 0.001  , true);
-      curAction.play();
-    }
-    
-  }
-
-  Exit() {
-  }
-
-  Update(timeElapsed, input) {
-      
-
-    this._parent.SetState('Idle1');
-  }
-};
-
-
-class IdleState1 extends State {
-  constructor(parent) {
-    super(parent);
-  }
-
-  get Name() {
-    return 'Idle1';
-  }
-
-  Enter(prevState) {
-    const idleAction = this._parent._proxy._animations['Idle1'].action;
-    if (prevState) {
-      const prevAction = this._parent._proxy._animations[prevState.Name].action;
-      idleAction.time = 0.0;
-      idleAction.enabled = true;
-      idleAction.setEffectiveTimeScale(1.0);
-      idleAction.setEffectiveWeight(1.0);
-      idleAction.crossFadeFrom(prevAction, 0.5, true);
-      idleAction.play();
-    } else {
-      idleAction.play();
-    }
-  }
-
-  Exit() {
-  }
-
-  Update(_, input) {
-    if (input._keys.forward1 || input._keys.backward1) {
-      this._parent.SetState('Run1');
-    } else if (input._keys.m) {
-      this._parent.SetState('Punch1');
-    }
-  }
-};
-
-//-------------------------------------------character 2 ---------------------------------------------------------------
-class JumpState2 extends State {
-  constructor(parent) {
-    super(parent);
-
-    this._FinishedCallback = () => {
-      this._Finished();
-    }
-  }
-
-  get Name() {
-    return 'Jump2';
-  }
-
-  Enter(prevState) {
-    const curAction2 = this._parent._proxy._animations['Jump2'].action;
-    const mixer = curAction2.getMixer();
-    mixer.addEventListener('finished', this._FinishedCallback);
-
-    if (prevState) {
-      const prevAction = this._parent._proxy._animations[prevState.Name].action;
-
-      curAction2.reset();  
-      curAction2.setLoop(THREE.LoopOnce, 1);
-      curAction2.clampWhenFinished = true;
-      curAction2.crossFadeFrom(prevAction, 0.2, true);
-      curAction2.play();
-    } else {
-      curAction2.play();
-    }
-  }
-
-  _Finished() {
-    this._Cleanup();
-    this._parent.SetState('Idle2');
-  }
-
-  _Cleanup() {
-    const action = this._parent._proxy._animations['Jump2'].action;
-    
-    action.getMixer().removeEventListener('finished', this._CleanupCallback);
-  }
-
-  Exit() {
-    this._Cleanup();
-  }
-
-  Update(_) {
-  }
-};
-
-
-class RunState2 extends State {
-  constructor(parent) {
-    super(parent);
-  }
-
-  get Name() {
-    return 'Run2';
-  }
-
-  Enter(prevState) {
-    const curAction2 = this._parent._proxy._animations['Run2'].action;
-    if (prevState) {
-      const prevAction = this._parent._proxy._animations[prevState.Name].action;
-
-      curAction2.enabled = true;
-      const ratio = curAction2.getClip().duration / prevAction.getClip().duration;
-      curAction2.time = prevAction.time * ratio;
-      curAction2.crossFadeFrom(prevAction, 0.5, true);
-      curAction2.play();
-    }
-  }
-
-  Exit() {
-  }
-
-  Update(timeElapsed, input) {
-    if (input._keys.forward2 || input._keys.backward2) {
-      if (input._keys.shift) {
-        this._parent.SetState('Jump2');
-      }
-      if (input._keys.enter) {
-        this._parent.SetState('Punch2');
-      }
-      return;
-    }
-
-    this._parent.SetState('Idle2');
-  }
-};
-
-
-class PunchState2 extends State {
-  constructor(parent) {
-    super(parent);
-  }
-
-  get Name() {
-    return 'Punch2';
-  }
-
-  Enter(prevState) {
-    const curAction = this._parent._proxy._animations['Punch2'].action;
-    if (prevState) {
-      const prevAction = this._parent._proxy._animations[prevState.Name].action;
-      curAction.enabled = true;
-      const ratio = curAction.getClip().duration / prevAction.getClip().duration;
-      curAction.time = 60;
-      curAction.crossFadeFrom(prevAction, 0.001  , true);
-      curAction.play();
-    }
-    
-  }
-
-  Exit() {
-  }
-
-  Update(timeElapsed, input) {
-      
-
-    this._parent.SetState('Idle2');
-  }
-};
-
-
-class IdleState2 extends State {
-  constructor(parent) {
-    super(parent);
-  }
-
-  get Name() {
-    return 'Idle2';
-  }
-
-  Enter(prevState) {
-    const idleAction = this._parent._proxy._animations['Idle2'].action;
-    if (prevState) {
-      const prevAction = this._parent._proxy._animations[prevState.Name].action;
-      idleAction.time = 0.0;
-      idleAction.enabled = true;
-      idleAction.setEffectiveTimeScale(1.0);
-      idleAction.setEffectiveWeight(1.0);
-      idleAction.crossFadeFrom(prevAction, 0.5, true);
-      idleAction.play();
-    } else {
-      idleAction.play();
-    }
-  }
-
-  Exit() {
-  }
-
-  Update(_, input) {
-    if (input._keys.forward2 || input._keys.backward2) {
-      this._parent.SetState('Run2');
-    } else if (input._keys.enter) {
-      this._parent.SetState('Punch2');
-    }
-  }
-};
-
-
-//--------------------------------------------------------------- Character Control -----------------------------------------------------------------------------------------------
-export default class BasicCharacterController {
-  constructor(params) {
-    this._Init(params);
-  }
-
-  //test commit2
-  //Make movements false 
-  _Init(params) {
-    this._params = params;
-   
-    //Set up the physics of the model moving
-    this._decceleration = new THREE.Vector3(-0.0005, -0.0001, -5.0);
-    this._acceleration = new THREE.Vector3(4, 1, 200.0);
-    this._velocity = new THREE.Vector3(0, 0, 0);
-
-    this._animations = {};
-    this._input = new BasicCharacterControllerInput();
-    this._stateMachine = new CharacterFSM(
-      new BasicCharacterControllerProxy(this._animations));
-    
-    this._LoadModel1();
-    this._LoadModel2();
-  }
-
-  _LoadModel1() {
-    const loader = new FBXLoader();
-    loader.setPath('./resources/Characters/');
-    loader.load('Ninja.fbx', (fbx) => {
-      fbx.scale.setScalar(0.1);
-      fbx.position.set(-25, 0, -27);
-      fbx.rotation.set(0, Math.PI / 2, 0)
-      fbx.traverse(c => {
-        c.castShadow = true;
-      });
-
-    
-      this._target = fbx;
-      this._params.scene.add(this._target);
-
-      this._mixer = new THREE.AnimationMixer(this._target);
-
-      this._manager = new THREE.LoadingManager();
-      this._manager.onLoad = () => {
-        this._stateMachine.SetState('Idle1');
-      };
-
-      const _OnLoad = (animName, anim) => {
-        const clip = anim.animations[0];
-        const action = this._mixer.clipAction(clip);
-  
-        this._animations[animName] = {
-          clip: clip,
-          action: action,
-        };
-      };
-
-      const loader = new FBXLoader(this._manager);
-      loader.setPath('./resources/Movements/');
-      loader.load('Run1.fbx', (a) => { _OnLoad('Run1', a); });
-      loader.load('Punch1.fbx', (a) => { _OnLoad('Punch1', a); });
-      loader.load('Idle1.fbx', (a) => { _OnLoad('Idle1', a); });
-      loader.load('Jump1.fbx', (a) => { _OnLoad('Jump1', a); });
-    });
-  }
-
-  _LoadModel2() {
-    const loader2 = new FBXLoader();
-    loader2.setPath('./resources/Characters/');
-    loader2.load('The Boss.fbx', (fbx) => {
-      fbx.scale.setScalar(0.1);
-      fbx.position.set(30, 0, 27);
-      fbx.rotation.set(0, -Math.PI / 1.5, 0)
-      fbx.traverse(c => {
-        c.castShadow = true;
-      });
-
-      this._target = fbx;
-      this._params.scene.add(this._target);
-
-      this._mixer = new THREE.AnimationMixer(this._target);
-
-      this._manager = new THREE.LoadingManager();
-      this._manager.onLoad = () => {
-        this._stateMachine.SetState('Idle2');
-      };
-
-      const _OnLoad = (animName, anim) => {
-        const clip = anim.animations[0];
-        const action = this._mixer.clipAction(clip);
-
-        this._animations[animName] = {
-          clip: clip,
-          action: action,
-        };
-      };
-
-      const loader2 = new FBXLoader(this._manager);
-      loader2.setPath('./resources/Movements/');
-      loader2.load('Run2.fbx', (a) => { _OnLoad('Run2', a); });
-      loader2.load('Punch2.fbx', (a) => { _OnLoad('Punch2', a); });
-      loader2.load('Idle2.fbx', (a) => { _OnLoad('Idle2', a); });
-      loader2.load('Jump2.fbx', (a) => { _OnLoad('Jump2', a); });
-    });
-    
-  }
-
-
-  
-
-  //Functions for the physics of the movements
-  Update(timeInSeconds) {
-    if (!this._target) {
-      return;
-    }
-
-    this._stateMachine.Update(timeInSeconds, this._input);
-
-    const velocity = this._velocity;
-    const frameDecceleration = new THREE.Vector3(
-        velocity.x * this._decceleration.x,
-        velocity.y * this._decceleration.y,
-        velocity.z * this._decceleration.z
-    );
-    frameDecceleration.multiplyScalar(timeInSeconds);
-    frameDecceleration.z = Math.sign(frameDecceleration.z) * Math.min(
-        Math.abs(frameDecceleration.z), Math.abs(velocity.z));
-
-    velocity.add(frameDecceleration);
-
-    const controlObject = this._target;
-    const _Q = new THREE.Quaternion();
-    const _A = new THREE.Vector3();
-    const _R = controlObject.quaternion.clone();
-
-    const acc = this._acceleration.clone();
-    if (this._input._keys.shift) {
-      acc.multiplyScalar(2.0);
-    }
-
-    if (this._input._keys.forward1) {
-      velocity.z += acc.z * timeInSeconds;
-    }
-    if (this._input._keys.backward1) {
-      velocity.z -= acc.z * timeInSeconds;
-    }
-    if (this._input._keys.left1) {
-      _A.set(0, 1, 0);
-      _Q.setFromAxisAngle(_A, 2.0 * Math.PI * timeInSeconds * this._acceleration.y);
-      _R.multiply(_Q);
-    }
-    if (this._input._keys.right1) {
-      _A.set(0, 1, 0);
-      _Q.setFromAxisAngle(_A, 2.0 * -Math.PI * timeInSeconds * this._acceleration.y);
-      _R.multiply(_Q);
-    }
-
-    controlObject.quaternion.copy(_R);
-
-    const oldPosition = new THREE.Vector3();
-    oldPosition.copy(controlObject.position);
-
-    const forward = new THREE.Vector3(0, 0, 1);
-    forward.applyQuaternion(controlObject.quaternion);
-    forward.normalize();
-
-    const sideways = new THREE.Vector3(1, 0, 0);
-    sideways.applyQuaternion(controlObject.quaternion);
-    sideways.normalize();
-
-    sideways.multiplyScalar(velocity.x * timeInSeconds);
-    forward.multiplyScalar(velocity.z * timeInSeconds);
-
-    controlObject.position.add(forward);
-    controlObject.position.add(sideways);
-
-    oldPosition.copy(controlObject.position);
-
-    if (this._mixer) {
-      this._mixer.update(timeInSeconds);
-    }
-  }
-}
-
-
+import BasicCharacterController from './Character_and_Movements.js'
 
 //---------------------------------------------------------------- Creating The World ----------------------------------------------------------------------------------------------
